Allows the creation of pre_trade_mgt in a strategy. The code is backwards compatible, a strategy without will be unaffected
pre_trade_mgt is called after buy/sell is populated and allows an opportunity to
cancel a trade for any given reason.

A sample TEST strategy is provided that
 - buys each even minute and sells each odd

pre_trade_mgt included in the strategy with code example that
tests if the last X of Y trades for a pair were profitable of not,
and cancel the buy if not.

Dynamic black listing.

This works in Dry and Live.
It is not linked to back test yet.